Backend

See merge request s10-bigdata-recom-sub2/S10P22A602!266

+++ FTActors-be/src/main/java/com/a602/actors/domain/montage/service/MontageFileService.java
@@ -1,28 +1,23 @@
 package com.a602.actors.domain.montage.service;
 
 
-import com.a602.actors.domain.member.repository.MemberRepository;
 import com.a602.actors.domain.montage.dto.MontageDto;
 import com.a602.actors.domain.montage.dto.MontageReportDto;
 import com.a602.actors.domain.montage.entity.Montage;
-import com.a602.actors.domain.montage.entity.Report;
 import com.a602.actors.domain.montage.repository.MontageRepository;
-
 import com.a602.actors.global.common.config.FileUtil;
 import com.a602.actors.global.common.enums.FolderType;
 import com.a602.actors.global.exception.ExceptionCodeSet;
 import com.a602.actors.global.exception.FileException;
 import com.a602.actors.global.exception.MontageException;
+import com.a602.actors.global.jwt.util.JWTUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 
 // S3 연결
 @Service
@@ -30,15 +25,18 @@ import java.util.Optional;
 public class MontageFileService {
 
     private final MontageRepository montageRepository;
+    private final JWTUtil jwtUtil;
 
-    public MontageFileService(MontageRepository montageRepository) {
+    public MontageFileService(MontageRepository montageRepository, JWTUtil jwtUtil) {
         this.montageRepository = montageRepository;
+        this.jwtUtil = jwtUtil;
     }
     public List<MontageDto.Montages> getAllMontageList(){
         return montageRepository.getAllMontages();
     }
 
-    public List<MontageDto.Montages> getMyMontage(Long memberId){
+    public List<MontageDto.Montages> getMyMontage(){
+        Long memberId = jwtUtil.getLoginMemberId();
         return montageRepository.getMyMontages(memberId).stream().map(MontageDto.Montages::toDto).toList();
     }
 
@@ -61,6 +59,13 @@ public class MontageFileService {
         Montage montage = montageRepository.getMontage(montageId)
                 .orElseThrow(() -> new MontageException(ExceptionCodeSet.ENTITY_NOT_EXISTS));
 
+        Long memberId = jwtUtil.getLoginMemberId();
+        
+        // 업로드 한 사람이 아니면
+        if(!Objects.equals(montage.getMember().getId(), memberId)){
+            throw new MontageException(ExceptionCodeSet.INVALID_AUTHORIZATION);
+        }
+
         FileUtil.deleteFile(montage.getSavedName(), FolderType.MONTAGE_PATH);
 
         montageRepository.deleteMontage(montageId);
@@ -69,7 +74,7 @@ public class MontageFileService {
     }
 
     public boolean pushLike(Long montageId){
-        Long memberId = 1L;
+        Long memberId = jwtUtil.getLoginMemberId();;
         return montageRepository.addLike(montageId, memberId);
     }
 
@@ -78,7 +83,7 @@ public class MontageFileService {
                          MultipartFile file,
                          Long montageId) throws IOException, MontageException, FileException {
 
-        Long reporterId = 1L;
+        Long reporterId = jwtUtil.getLoginMemberId();
 
         // 파일이 없는 경우
         if(file == null){
@@ -100,7 +105,7 @@ public class MontageFileService {
 
         String url = FileUtil.uploadFile(file, savedName, FolderType.REPORT_PATH);
         log.info("URL : {}", url);
-        montageRepository.addReport(1L, montageId, req.getReason(), url);
+        montageRepository.addReport(reporterId, montageId, req.getReason(), url);
 
         return "";
     }
